xmerl: Use context namespace declarations to resolve prefix node tests
Previously, a match would not be found if the namespace prefix in the
XPath query was not contained in the original document.  This allows
the `namespace' option to provide a prefix that will be resolved to a
namespace URI.

See Section 2.3 of the XPath 1.0 specification for the behaviour of
'NCName:*' node tests.